fix(utils): make our own capitalize helper
ember/string is deprecated and also an extreme pain in the neck
cause it used to be a "fake" package that was actually shipped with
ember core. That is no longer the case starting from some unknown ember
version, so now a "real" dependency on it is required.

However it is also deprecated, so the best approach is to just avoid it.

Code for capitalization taken from ember/string to ensure same behavior.