fix(types): add back ember cli, but as a devDep
Adding it as a dep is no longer needed since the compilation is handled
by babel.
Keeping it as a dep anyway somehow breaks the use of yalc (unknown why
this is)
Removing the dep altogether gets rid of the very useful precompile and
clean commands that are not easy to replicate.
So moving it to devDeps gives us the best of both worlds: working yalc,
working compilation, and working prepack and postpack scripts